Leave all testing to the testing role, not the setup role
The testing role includes a large number of test cases designed to fail in specific ways when specific problems arise. When one-off tests are included in the main setup playbook, they can short-circuit this process, resulting in an indication that something has gone wrong, but without the proper context to indicate exactly what has gone wrong.